GKI: acpm_ipc: fix SRAM memory prefetch issue with strstr
The strstr function is causing a page fault in memcpy. This is expected
because the address is an IO address (from sram). According to the
device-io documentation, "Do not use memset or memcpy on IO addresses;
they are not guaranteed to copy data in order." Fix this to use
memcpy_fromio to compare the string instead.

Signed-off-by: Will McVicker <willmcvicker@google.com>
Change-Id: I3d2925a56c33181236bbd6f6e86152daccb1c5c8
(cherry picked from commit 3d8ecd6dd898c5832fbbdbd3cb82eb426c85bffa)
Signed-off-by: Panchajanya1999 <kernel@panchajanya.dev>