gs_tmu_v2: Assign maximum RT priority to IRQ worker thread
If the TMU worker doesn't complete on time, then the CPU can crash due to
not being throttled quickly enough. This is especially prevalent on the
IPOP silicon variant, where not throttling the CPU can easily make it crash
while running a benchmark.

In order to be extra safe, set the IRQ worker to maximum RT priority to
ensure stability of the CPU.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>